+++ packages/@lou.codes/cron/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## 2.2.1
+
+### Patch Changes
+
+-   ⬆️ upgrade `functional-expression` to `v2.0.0`.
+
 ## 2.2.0
 
 ### Minor Changes